Native Libraries: moved code to JNIUtils.cpp and *MessageDialog.cpp
