+++ flatlaf-natives/flatlaf-natives-linux/src/main/headers/JNIUtils.h
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2025 FormDev Software GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include <gtk/gtk.h>
+#include <jni.h>
+
+/**
+ * @author Karl Tauber
+ */
+
+//---- class AutoReleaseStringUTF8 --------------------------------------------
+
+class AutoReleaseStringUTF8 {
+	JNIEnv* env;
+	jstring javaString;
+	const char* chars;
+
+public:
+	AutoReleaseStringUTF8( JNIEnv* _env, jstring _javaString );
+	~AutoReleaseStringUTF8();
+
+	operator const gchar*() { return chars; }
+};
